moved files and added banner image
