+++ src/components/App.js
@@ -1,10 +1,12 @@
 import React from 'react';
-import Header from './header/Header';
+import Header from './Header';
+import Banner from './Banner';
 
 const App = () => {
   return (
     <div className="ui container">
       <Header/>
+      <Banner/>
     </div>
   );
 };
